fix: redone the reviewer's comments
